Moving Struct to own file + extending additional case classes
